extract more beacon state utils
