+++ res/css/components/views/beacon/_RoomLiveShareWarning.scss
@@ -48,6 +48,10 @@ limitations under the License.
     margin-left: $spacing-16;
 }
 
+.mx_RoomLiveShareWarning_stopButton {
+    margin-left: $spacing-16;
+}
+
 .mx_RoomLiveShareWarning_closeButtonIcon {
     height: $font-18px;
     padding: $spacing-4;
