+++ src/utils/beacon/index.ts
@@ -17,3 +17,4 @@ limitations under the License.
 export * from './duration';
 export * from './geolocation';
 export * from './useBeacon';
+export * from './useOwnLiveBeacons';
